Fix incremental build Java generation.
Translate task is requested to put generated source into a target
directory "main", however dafny actually places it into a directory
callled "main-java" and the translate task performs a rename at the
end to compensate. The problem is this rename can fail, for example
on an incremental build, when the target directory already exists.

Instead keep "main-java" as target and pass this modified target
onto subsequent gradle tasks as new source to be compiled.